backend auth.ts refactor complete
